requirements.txt
dj-database-url==0.5.0
Django==2.2.4
django-admin-list-filter-dropdown==1.0.2
django-adminfilters==1.1.0
django-advanced-filters==1.1.1
django-appconf==1.0.3
django-braces==1.13.0
django-compressor==2.3
django-libsass==0.7
djangorestframework==3.9.3
libsass==0.19.2
mod-wsgi==4.6.7
mysqlclient==1.4.4
PyMySQL==0.9.3
pytz==2019.2
rcssmin==1.0.6
rjsmin==1.1.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
